Explicitly break Python 2 support in all our Python libraries https://bugs.webkit.org/show_bug.cgi?id=260922
Reviewed by NOBODY (OOPS!).

This makes the following packages raise ImportError on Python < 3.6,
thereby breaking support for Python 2. Nothing should be relying on
this any more, and this provides us with a path towards cleaning up a
lot of legacy support.

Bot-watchers: If you're reading this, I'm sorry. I hope I've not
completely set everything on fire. Please file a bug for whatever this
broke and just revert this ASAP and ensure any services that regressed
get redeployed.

* Tools/Scripts/libraries/reporelaypy/reporelaypy/__init__.py:
* Tools/Scripts/libraries/resultsdbpy/resultsdbpy/__init__.py:
* Tools/Scripts/libraries/webkitbugspy/webkitbugspy/__init__.py:
* Tools/Scripts/libraries/webkitcorepy/webkitcorepy/__init__.py:
* Tools/Scripts/libraries/webkitflaskpy/webkitflaskpy/__init__.py:
* Tools/Scripts/libraries/webkitscmpy/webkitscmpy/__init__.py:
* Tools/Scripts/webkitpy/__init__.py: