[WebDriver][socket] Titles containing multibyte characters cannot be retrieved correctly https://bugs.webkit.org/show_bug.cgi?id=270063
Reviewed by NOBODY (OOPS!).

If the document title contains multibyte characters such as Japanese
or entity references, the "Get Title" result will be garbled.

The title string is obtained by JavaScript's "document.title()"
and is UTF-8 encoded in WebDriverService::sendResponse.

However, when formatting the result as an http message,
StringBuilder.append() uses fromLatin1() to generate a WTF::String
from the byte data.
This seems to be causing the multibyte characters to be garbled.

This patch changes to use String::fromUTF8() before concatenation to
restore the correct WTF::String even if it contains multibyte characters.

* Source/WebDriver/socket/HTTPServerSocket.cpp: