Don't use RegisterPair for Int32 values
We signal that an ArgumentLocation is 32-bit by attaching the usedWidth
(only on 32-bits). This means we need to interpose code to create a
B3::ValueRep out of payloadGPR only, wherever we instantiate a ValueRep
out of the ArgumentLocation.

This also requires threading valueSize through to where we create the
ArgumentLocation in WasmCallingConvention.h.

Another approach would be to set the tagGPR to InvalidGPRReg in the
JSValueRegs of the ArgumentLocation, but there's code somewhere that
depends on the tagGPR being present, which I haven't been able to track
down.