[Site Isolation] Make performDragOperation aware of site-isolated i-frames https://bugs.webkit.org/show_bug.cgi?id=273847 rdar://127699330
Reviewed by NOBODY (OOPS!).

This PR starts to handle performDragOperation for site-isolated i-frames. This is
how files that are drag and dropped make it into a webpage. It also prevents dragging
a file onto a page that handles these events from instead navigating the page, which
is the behavior dragging a file into a site-isolated i-frame before this change.

Added an API test to exercise this behavior.

* Source/WebCore/page/DragController.cpp:
(WebCore::DragController::performDragOperation):
* Source/WebCore/page/DragController.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::performDragOperation):
* Source/WebKit/UIProcess/WebPageProxy.h:
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::remoteEventDataForPosition):
(WebKit::WebPage::performDragOperation):
* Source/WebKit/WebProcess/WebPage/WebPage.h:
* Source/WebKit/WebProcess/WebPage/WebPage.messages.in:
* Source/WebKitLegacy/mac/WebView/WebView.mm:
(-[WebView performDragOperation:]):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::TEST(SiteIsolation, DragDrop)):