WASM tail calls should work in OMG tier https://bugs.webkit.org/show_bug.cgi?id=273875
Reviewed by NOBODY (OOPS!).

Tom's original patch was correct, but we told him to elide moves to temporary storage when they
weren't needed. Unfortunately, the current version on ToT clobbers a bunch of important stuff. Let's
fix that.

We also fix up LLInt, since it had a bit of bit rot.

Finally, we add a new kind of tail call, a fake tail call, for the case when an inlined call makes
a tail call.

There is still more work to be done to ensure that this code is correct; Tests for stack traces,
inlined calls that clobber instance, and more advanced stack / regalloc situations are all
still needed.

* JSTests/wasm/stress/simd-tail-call-simple.js: Added.
* JSTests/wasm/stress/tail-call-js.js: Added.
* JSTests/wasm/stress/tail-call-simple.js: Added.
* Source/JavaScriptCore/wasm/WasmOMGIRGenerator.cpp: