[GStreamer][MSE] SleepDisabler not reliably created when playing YouTube videos https://bugs.webkit.org/show_bug.cgi?id=219354
Reviewed by NOBODY (OOPS!).

When playing videos the display screensaver should be disabled, and when the video is paused it
should be re-enabled. This is mostly useful on desktop systems. The sleep disabler is handled by
HTMLMediaElement.

Until this patch display sleep wasn't consistently disabled as a result of playback request. By
notifying the MediaPlayer of successful GStreamer pipeline state changes the HTMLMediaElement gets
another opportunity to update the sleep disabling state. This is done already in the
MediaPlayerPrivateGStreamer::updateStates() method, but for some reason the MSE player has a
different implementation of that method.

Test: media/media-source/media-source-video-play-holds-sleep-assertion.html

* LayoutTests/media/media-source/media-source-video-play-holds-sleep-assertion-expected.txt: Added.
* LayoutTests/media/media-source/media-source-video-play-holds-sleep-assertion.html: Added.
* LayoutTests/platform/mac-wk1/TestExpectations:
* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebCore/html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::progressEventTimerFired):
* Source/WebCore/platform/graphics/gstreamer/MediaPlayerPrivateGStreamer.h:
* Source/WebCore/platform/graphics/gstreamer/mse/MediaPlayerPrivateGStreamerMSE.cpp:
(WebCore::MediaPlayerPrivateGStreamerMSE::pause):
(WebCore::MediaPlayerPrivateGStreamerMSE::checkPlayingConsistency):
(WebCore::MediaPlayerPrivateGStreamerMSE::setShouldDisableSleep):
(WebCore::MediaPlayerPrivateGStreamerMSE::updateStates):
(WebCore::MediaPlayerPrivateGStreamerMSE::timeIsProgressing const):
* Source/WebCore/platform/graphics/gstreamer/mse/MediaPlayerPrivateGStreamerMSE.h: