[WebDriver][socket] Add confirmation of socket connection before sending commands to the backend https://bugs.webkit.org/show_bug.cgi?id=274509
Reviewed by NOBODY (OOPS!).

The SessionHost of WebDriver is trying to send a command to backend
without checking whether the socket is connected or not.
If the socket is disconnected, the completionHandler will remain added
to the queue and will not be called by anyone, and the process will not
be completed.
To solve this problem, add code to check the status of the socket
before sending the command.

* Source/WebDriver/SessionHost.cpp:
(WebDriver::SessionHost::sendCommandToBackend):