[WGSL] shader,execution,expression,constructor,non_zero:* is still failing https://bugs.webkit.org/show_bug.cgi?id=274600 rdar://128624933
Reviewed by NOBODY (OOPS!).

I fixed the tests in 278991@main, but didn't update the expectations as it required
updating the CTS tests first, and the tests got broken before I could update it.
The fix was fairly simple, we just weren't generating the PackedVec3 struct for a
couple tests that needed it, which was fixed by changing where we detect that we
need to generate the struct.

* LayoutTests/http/tests/webgpu/webgpu/shader/execution/expression/constructor/non_zero-expected.txt: Added.
* Source/WebGPU/WGSL/GlobalVariableRewriter.cpp:
(WGSL::RewriteGlobalVariables::pack):
(WGSL::RewriteGlobalVariables::packArrayResource):
(WGSL::RewriteGlobalVariables::packArrayType):