REGRESSION: Bad flashing scrolling msn.com article when search field has focus (content-visibility) https://bugs.webkit.org/show_bug.cgi?id=274624 rdar://127743319
Reviewed by NOBODY (OOPS!).

The cause of the flashing is that the IntersectionObserver computed that the content-visibility: auto elements
were going in and out of the visual viewport rapidly. This was due to the visual viewport’s value changing
between two locations, due to the difference in how it is computed between LocalFrameView::updateLayoutViewport
and AsyncScrollingCoordinator::reconcileScrollingState. In LocalFrameView::updateLayoutViewport, the visual viewport is used
in computeLayoutViewportOrigin to compute the new location of the layout viewport. When the keyboard is
visible, m_visualViewportOverrideRect is sent via WebPage::updateVisibleContentRects, but is not udpated during
unstable state, causing computeLayoutViewportOrigin to use a stale version of the visual viewport. The solution
is to always update the computeLayoutViewportOrigin value if received from the ui process.

* Source/WebKit/WebProcess/WebPage/ios/WebPageIOS.mm:
(WebKit::WebPage::updateVisibleContentRects):