Web Inspector: At some narrow-ish widths objects and disclosure triangles in console output disappear rdar://121861970 https://bugs.webkit.org/show_bug.cgi?id=274630
The object display, including the disclosure triangles, are instances
of ObjectTreeBaseTreeElement's subclasses, which have the
`.object-tree` class name as DOM elements. This commit improves the
styles of these elements by using them always `display: inline` since
it worked best with the text truncating feature that the elements have.

While making the `.object-tree` elements `display: inline`, the tree
element was rendered incorrectly when displayed as an indexed item,
with a preceding array index or a label like "key" or "value".
The fix is to use a `display: flex` for the entry row instead, so
the `.index-name` (the label) and the `.index-value` (containing the
`.object-tree`) are always side by side instead of the `.index-value`
being wrapped and shown on a second line.

* Source/WebInspectorUI/UserInterface/Views/ObjectTreeView.css:
(.object-tree):
* Source/WebInspectorUI/UserInterface/Views/ConsoleMessageView.css:
(.console-top-level-message .object-tree): Deleted.
(.console-top-level-message .object-tree .object-tree): Deleted.
  - Always use `display: inline` for the object tree element. It works
    well with the text-truncating styles the element has.

* Source/WebInspectorUI/UserInterface/Views/ObjectTreeMapEntryTreeElement.css:
(.object-tree-array-index.object-tree-map-entry > .titles > .title):
  - Use `display: flex` so that the name and the value elements are
    always side by side instead of the value wrapped to a second line
    when it gets wide.

(.object-tree-array-index.object-tree-map-entry > .titles > .title > .index-name):
  - Use `min-width` instead of `width` to go with the `display: flex`
    in the parent.

(.object-tree-array-index.object-tree-map-entry > .titles > .title > .index-value):
  - Technically not needed for this fix, but without this change when
    window gets narrow, there are no ellipses at the end of the object
    preview elements that are now cut off. This change simply make
    ellipses show up so the UI is more consistent with ellipses already
    showing up elsewhere that an object tree shows up. (But not console
    outputs that are strings, as filed in https://webkit.org/b/270992.)